Make sure conditional syntax in darrc has actual results.
It turns out that the way dar deals with the darrc file is it
determines what mode it's in, then pays attention to the switches
listed under that mode in the darrc file. This means that if you
reference the darrc file on a switch on the dar command line that
comes before any switches that let dar know which mode it's going
to be running in (-c -t -x etc), dar will not pay attention to any
of the conditional switches.