add a couple of scripts to deal with expected sizes of par files.
PAR files are slightly larger than the redundancy data they contain, because of the checksums and headers involved. I want to get a rough idea of those sizes ahead of time, but being a new father have no attention span to work it out from the PAR source. So, empirical discovery it is.