Add support for fetching and recovering less than a full disc set of files.
The _obtain_recovery_set_files method used to just get all the files from a
set, and run all the parchive commands to make sure all the files were valid.
Now, the _fetch_some_slices method can get some slices, maybe not at the
beginning of a set, and run however many parchive commands are needed on them.
The way it is called now, the only difference in dar's behavior is that it only
runs parchive once when getting the last slice, rather than recovering the
entire last disc set. But better changes are coming.